velora/backend/requirements.txt
 fastapi==0.115.2
 uvicorn[standard]==0.31.1
 sqlalchemy==2.0.31
psycopg[binary]==3.2.3
 python-socketio==5.11.0
 aiohttp==3.9.5
 numpy==1.26.4
 scipy==1.13.1
 pandas==2.2.3
 torch==2.4.1
 numba==0.59.1
 ujson==5.10.0
 pydantic==2.9.2
 python-dotenv==1.0.1
 pytest==8.2.0
 httpx==0.27.2
 pydantic-settings==2.4.0


